alpaca-producer/requirements.txt
websocket-client==1.7.0
boto3==1.34.18
python-dotenv==1.0.0

